feat(stratum_v1): suggest fractional difficulty below 1
Compute suggested difficulty as a float and let it fall below 1 instead
of clamping to a minimum of 1. A very slow worker such as the CPU miner
needs a target it can actually meet. Clamping gave it difficulty 1,
which it rarely produces shares for.

The value stays an f64 through the suggest path, and the client decides
the wire encoding. Whole difficulties still serialize as integers, the
form pools expect, and only values below 1 go out as a fraction.
Production boards run far above 1 and are unaffected.

+++ mujina-miner/src/stratum_v1/messages.rs
@@ -77,8 +77,9 @@ pub enum ClientCommand {
     /// Submit a share to the pool
     SubmitShare(SubmitParams),
 
-    /// Suggest a new difficulty to the pool
-    SuggestDifficulty(u64),
+    /// Suggest a new difficulty to the pool. Fractional below 1 (for very
+    /// slow workers like the CPU miner); sent as an integer at or above 1.
+    SuggestDifficulty(f64),
 }
 
 /// Mining job notification from pool (mining.notify).
